+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -20,7 +20,7 @@
 
 20200917:
 3: 0011, 0015, 0016, 1291
-4: 
+4: 0145
 5: 0010
 
 20200916:
