+++ src/main/java/com/zea7ot/leetcode/lvl4/lc0145/SolutionApproach1ReversingPreorderIterative.java
@@ -10,28 +10,34 @@
  */
 package com.zea7ot.leetcode.lvl4.lc0145;
 
+import java.util.ArrayDeque;
+import java.util.Deque;
 import java.util.LinkedList;
 import java.util.List;
 
 import com.zea7ot.utils.data_structure.tree.TreeNode;
 
-public class SolutionApproach0ReversePostorderRecursive {
+public class SolutionApproach1ReversingPreorderIterative {
     public List<Integer> postorderTraversal(TreeNode root) {
         LinkedList<Integer> ans = new LinkedList<>();
         // sanity check
         if (root == null)
             return ans;
 
-        reversePostorder(root, ans);
+        Deque<TreeNode> stack = new ArrayDeque<>();
+        stack.push(root);
+
+        while (!stack.isEmpty()) {
+            TreeNode top = stack.pop();
+            ans.addFirst(top.val);
+
+            if (top.left != null)
+                stack.push(top.left);
+
+            if (top.right != null)
+                stack.push(top.right);
+        }
+
         return ans;
-    }
-
-    private void reversePostorder(TreeNode node, LinkedList<Integer> res) {
-        if (node == null)
-            return;
-
-        res.addFirst(node.val);
-        reversePostorder(node.right, res);
-        reversePostorder(node.left, res);
     }
 }
